@@ -1,61 +1,63 @@
 {
   "name": "HiveStore.WebApp",
   "version": "0.0.0",
-  "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --extract-css",
-    "build": "ng build --extract-css",
-    "build:ssr": "npm run build -- --app=ssr --output-hashing=media",
+    "start": "ng serve",
+    "build": "ng build",
+    "build:ssr": "ng run HiveStore.WebApp:server:dev",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/cdk": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/material": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/platform-server": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "@nguniversal/module-map-ngfactory-loader": "^5.0.0-beta.5",
+    "@angular/animations": "6.1.10",
+    "@angular/cdk": "^6.4.2",
+    "@angular/common": "6.1.10",
+    "@angular/compiler": "6.1.10",
+    "@angular/core": "6.1.10",
+    "@angular/forms": "6.1.10",
+    "@angular/http": "6.1.10",
+    "@angular/material": "^6.4.2",
+    "@angular/platform-browser": "6.1.10",
+    "@angular/platform-browser-dynamic": "6.1.10",
+    "@angular/platform-server": "6.1.10",
+    "@angular/router": "6.1.10",
+    "@nguniversal/module-map-ngfactory-loader": "6.0.0",
+    "core-js": "^2.5.4",
+    "rxjs": "^6.0.0",
+    "zone.js": "^0.8.26",
     "aspnet-prerendering": "^3.0.1",
     "bootstrap": "^3.3.7",
-    "core-js": "^2.4.1",
-    "dom-to-image": "^2.6.0",
+    "jquery": "3.3.1",
+    "popper.js": "^1.14.3",
     "font-awesome": "^4.7.0",
-    "formdata-polyfill": "^3.0.10",
-    "html2canvas": "^1.0.0-alpha.12",
-    "jspdf": "^1.3.5",
-    "primeng": "^5.2.4",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
+    "primeng": "^6.0.0",
+    "primeicons": "^1.0.0-beta.9"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.0",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
+    "@angular-devkit/build-angular": "~0.6.0",
+    "@angular/cli": "~6.0.0",
+    "@angular/compiler-cli": "6.1.10",
+    "@angular/language-service": "^6.0.0",
+    "@types/jasmine": "~2.8.6",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.2.1",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "^3.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "~1.4.2",
+    "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "~2.7.2"
+  },
+  "optionalDependencies": {
+    "node-sass": "^4.9.3",
+    "protractor": "~5.4.0",
+    "ts-node": "~5.0.1",
+    "tslint": "~5.9.1"
   }
 }
