@@ -1,8 +1,6 @@
-import { HttpClient, HttpHeaders, HttpErrorResponse } from '@angular/common/http';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/observable/of';
-import { Observable } from 'rxjs/Observable';
+import { HttpHeaders, HttpErrorResponse } from '@angular/common/http';
+import { map, catchError } from 'rxjs/operators';
+import { Observable } from 'rxjs'
 import { AppSettings } from '../app.settings';
 import { BaseResponse } from '../models/base-response.model';
 import { SignIn } from '../models/sign-in.model';
@@ -16,24 +14,27 @@ export class AccountService extends BaseService {
       headers: new HttpHeaders({ 'Content-Type': 'application/json', 'Accept': 'application/json' })
     };
     const url = AppSettings.SIGN_IN_URL;
-    return this.http.post<BaseResponse>(url, singInDTO, httpOptions)
-      .map((res: BaseResponse) => this.extractData(res))
-      .catch((res: HttpErrorResponse) => this.handleError(res));
+    return this.http.post<BaseResponse>(url, singInDTO, httpOptions).pipe(
+      map((res: BaseResponse) => this.extractData(res)),
+      catchError((res: HttpErrorResponse) => this.handleError(res))
+    );
   }
 
   signOut(): Observable<string> {
     const url = AppSettings.SIGN_OUT_URL;
 
-    return this.http.get<BaseResponse>(url)
-      .map((res: BaseResponse) => this.extractData(res))
-      .catch((res: HttpErrorResponse) => this.handleError(res));
+    return this.http.get<BaseResponse>(url).pipe(
+      map((res: BaseResponse) => this.extractData(res)),
+      catchError((res: HttpErrorResponse) => this.handleError(res))
+    );
   }
 
   getSignedInUser(): Observable<User> {
     const url = AppSettings.GET_SIGNED_IN_USER;
 
-    return this.http.get<BaseResponse>(url)
-      .map((res: BaseResponse) => this.extractData(res))
-      .catch((res: HttpErrorResponse) => this.handleError(res));
+    return this.http.get<BaseResponse>(url).pipe(
+      map((res: BaseResponse) => this.extractData(res)),
+      catchError((res: HttpErrorResponse) => this.handleError(res))
+    );
   }
 }
