@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
-import { Observable } from 'rxjs/Observable';
+import { map, catchError, } from 'rxjs/operators';
+import { Observable, of } from 'rxjs'
 import { AccountService } from './account.service';
 
 @Injectable()
@@ -9,34 +10,19 @@ export class AuthGuard implements CanActivate {
   message: string;
 
   constructor(private accountService: AccountService) {
-
   }
 
   canActivate(
     next: ActivatedRouteSnapshot,
     state: RouterStateSnapshot): Observable<boolean> | Promise<boolean> | boolean {
-    return this.accountService.getSignedInUser()
-      .map(
-        (user) => {
-          console.log(user);
-          return true;
-        },
-        error => {
-          this.message = <any>error;
-          return false;
-        });
-  }
-
-  async getSignedInUser() {
-    let isUserAuthorized = false;
-    await this.accountService.getSignedInUser()
-      .subscribe(
-        (user) => {
-          isUserAuthorized = true;
-        },
-        error => {
-          this.message = <any>error;
-        });
-    return isUserAuthorized;
+    return this.accountService.getSignedInUser().pipe(
+      map((user) => {
+        console.log(user);
+        return true;
+      }),
+      catchError((error) => {
+        console.log(error);
+        return of(false);
+      }));
   }
 }
