@@ -1,8 +1,6 @@
-import { HttpClient, HttpHeaders, HttpErrorResponse } from '@angular/common/http';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/observable/of';
-import { Observable } from 'rxjs/Observable';
+import { HttpErrorResponse } from '@angular/common/http';
+import { map, catchError } from 'rxjs/operators';
+import { Observable } from 'rxjs'
 import { AppSettings } from '../app.settings';
 import { BaseResponse } from '../models/base-response.model';
 import { RequestInfo } from '../models/request-info.model';
@@ -13,8 +11,9 @@ export class InfrastructureService extends BaseService {
   getRequestInfo(): Observable<RequestInfo> {
     const url = AppSettings.GET_SERVER_INFO;
 
-    return this.http.get<BaseResponse>(url)
-      .map((res: BaseResponse) => this.extractData(res))
-      .catch((res: HttpErrorResponse) => this.handleError(res));
+    return this.http.get<BaseResponse>(url).pipe(
+      map((res: BaseResponse) => this.extractData(res)),
+      catchError((res: HttpErrorResponse) => this.handleError(res))
+    );
   }
 }
