@@ -1,8 +1,6 @@
-import { HttpClient, HttpHeaders, HttpErrorResponse } from '@angular/common/http';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/observable/of';
-import { Observable } from 'rxjs/Observable';
+import { HttpHeaders, HttpErrorResponse } from '@angular/common/http';
+import { map, catchError } from 'rxjs/operators';
+import { Observable } from 'rxjs';
 import { AppSettings } from '../app.settings';
 import { BaseResponse } from '../models/base-response.model';
 import { Role } from '../models/role.model';
@@ -12,9 +10,10 @@ export class RoleService extends BaseService {
 
   getAllRoles(): Observable<[Role]> {
     const url = AppSettings.GET_ALL_ROLES_URL;
-    return this.http.get<BaseResponse>(url)
-      .map((res: BaseResponse) => this.extractData(res))
-      .catch((res: HttpErrorResponse) => this.handleError(res));
+    return this.http.get<BaseResponse>(url).pipe(
+      map((res: BaseResponse) => this.extractData(res)),
+      catchError((res: HttpErrorResponse) => this.handleError(res))
+    );
   }
 
   saveRole(role: Role): Observable<string> {
@@ -22,8 +21,9 @@ export class RoleService extends BaseService {
       headers: new HttpHeaders({ 'Content-Type': 'application/json', 'Accept': 'application/json' })
     };
     const url = AppSettings.SAVE_ROLE_URL;
-    return this.http.post<BaseResponse>(url, role, httpOptions)
-      .map((res: BaseResponse) => this.extractData(res))
-      .catch((res: HttpErrorResponse) => this.handleError(res));
+    return this.http.post<BaseResponse>(url, role, httpOptions).pipe(
+      map((res: BaseResponse) => this.extractData(res)),
+      catchError((res: HttpErrorResponse) => this.handleError(res))
+    );
   }
 }
