@@ -10,7 +10,7 @@ using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 namespace HiveStore.DataContext.Migrations
 {
     [DbContext(typeof(HiveDataContext))]
-    [Migration("20181203041909_InitialCreate")]
+    [Migration("20181205091756_InitialCreate")]
     partial class InitialCreate
     {
         protected override void BuildTargetModel(ModelBuilder modelBuilder)
@@ -27,7 +27,8 @@ namespace HiveStore.DataContext.Migrations
                         .ValueGeneratedOnAdd()
                         .HasColumnName("USER_ID");
 
-                    b.Property<int>("AccessFailedCount");
+                    b.Property<int>("AccessFailedCount")
+                        .HasColumnName("ACCESS_FAILED_COUNT");
 
                     b.Property<string>("Address")
                         .HasColumnName("ADDRESS")
@@ -38,7 +39,8 @@ namespace HiveStore.DataContext.Migrations
                         .HasMaxLength(200);
 
                     b.Property<string>("ConcurrencyStamp")
-                        .IsConcurrencyToken();
+                        .IsConcurrencyToken()
+                        .HasColumnName("CONCURRENCY_STAMP");
 
                     b.Property<string>("Country")
                         .HasColumnName("COUNTRY")
@@ -53,9 +55,11 @@ namespace HiveStore.DataContext.Migrations
                         .HasColumnName("CREATED_DATE");
 
                     b.Property<string>("Email")
+                        .HasColumnName("EMAIL")
                         .HasMaxLength(256);
 
-                    b.Property<bool>("EmailConfirmed");
+                    b.Property<bool>("EmailConfirmed")
+                        .HasColumnName("EMAIL_CONFIRMED");
 
                     b.Property<string>("FirstName")
                         .IsRequired()
@@ -70,9 +74,11 @@ namespace HiveStore.DataContext.Migrations
                         .HasColumnName("LAST_NAME")
                         .HasMaxLength(200);
 
-                    b.Property<bool>("LockoutEnabled");
+                    b.Property<bool>("LockoutEnabled")
+                        .HasColumnName("LOCK_OUT_ENABLED");
 
-                    b.Property<DateTimeOffset?>("LockoutEnd");
+                    b.Property<DateTimeOffset?>("LockoutEnd")
+                        .HasColumnName("LOCK_OUT_END");
 
                     b.Property<string>("ModifiedBy")
                         .IsRequired()
@@ -83,22 +89,30 @@ namespace HiveStore.DataContext.Migrations
                         .HasColumnName("MODIFIED_DATE");
 
                     b.Property<string>("NormalizedEmail")
+                        .HasColumnName("NORMALIZED_EMAIL")
                         .HasMaxLength(256);
 
                     b.Property<string>("NormalizedUserName")
+                        .HasColumnName("NORMALIZED_USER_NAME")
                         .HasMaxLength(256);
 
-                    b.Property<string>("PasswordHash");
+                    b.Property<string>("PasswordHash")
+                        .HasColumnName("PASSWORD_HASH");
 
-                    b.Property<string>("PhoneNumber");
+                    b.Property<string>("PhoneNumber")
+                        .HasColumnName("PHONE_NUMBER");
 
-                    b.Property<bool>("PhoneNumberConfirmed");
+                    b.Property<bool>("PhoneNumberConfirmed")
+                        .HasColumnName("PHONE_NUMBER_CONFIRMED");
 
-                    b.Property<string>("SecurityStamp");
+                    b.Property<string>("SecurityStamp")
+                        .HasColumnName("SECURITY_STAMP");
 
-                    b.Property<bool>("TwoFactorEnabled");
+                    b.Property<bool>("TwoFactorEnabled")
+                        .HasColumnName("TWO_FACTOR_ENABLED");
 
                     b.Property<string>("UserName")
+                        .HasColumnName("USER_NAME")
                         .HasMaxLength(256);
 
                     b.HasKey("Id");
@@ -109,7 +123,7 @@ namespace HiveStore.DataContext.Migrations
                     b.HasIndex("NormalizedUserName")
                         .IsUnique()
                         .HasName("UserNameIndex")
-                        .HasFilter("[NormalizedUserName] IS NOT NULL");
+                        .HasFilter("[NORMALIZED_USER_NAME] IS NOT NULL");
 
                     b.ToTable("AspNetUsers");
                 });
